complete test fixes
