+++ eventkit_cloud/ui/static/ui/app/tests/actions/datacartActions.spec.js
@@ -2,7 +2,6 @@
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import axios from 'axios';
-import sinon from 'sinon';
 import MockAdapter from 'axios-mock-adapter';
 import * as actions from '../../actions/datacartActions';
 
@@ -118,20 +117,6 @@ describe('export actions', () => {
         });
     });
 
-    it('stepperNextDisabled should return MAKE_STEPPER_INACTIVE and false', () => {
-        expect(actions.stepperNextDisabled()).toEqual({
-            type: 'MAKE_STEPPER_INACTIVE',
-            stepperNextEnabled: false,
-        });
-    });
-
-    it('stepperNextEnabled should return MAKE_STEPPER_ACTIVE and true', () => {
-        expect(actions.stepperNextEnabled()).toEqual({
-            type: 'MAKE_STEPPER_ACTIVE',
-            stepperNextEnabled: true,
-        });
-    });
-
     it('submitJob should post job data', () => {
         const mock = new MockAdapter(axios, { delayResponse: 1 });
 
@@ -166,35 +151,6 @@ describe('export actions', () => {
             });
     });
 
-    it('getFormats should return formats from the api', () => {
-        const mock = new MockAdapter(axios, { delayResponse: 1 });
-        mock.onGet('/api/formats').reply(200, ['my formats']);
-
-        const expectedActions = [
-            { type: actions.types.GETTING_FORMATS },
-            { type: actions.types.FORMATS_RECEIVED, formats: ['my formats'] },
-        ];
-        const store = mockStore({});
-        return store.dispatch(actions.getFormats())
-            .then(() => {
-                expect(store.getActions()).toEqual(expectedActions);
-            });
-    });
-
-    it('getFormats should handle errors', () => {
-        const mock = new MockAdapter(axios, { delayResponse: 1 });
-        mock.onGet('/api/formats').reply(400);
-
-        const expectedActions = [
-            { type: actions.types.GETTING_FORMATS },
-        ];
-        const store = mockStore({});
-        return store.dispatch(actions.getFormats())
-            .then(() => {
-                expect(store.getActions()).toEqual(expectedActions);
-            });
-    });
-
     it('clearAoiInfo should return type CLEAR_AOI_INFO and no action', () => {
         expect(actions.clearAoiInfo()).toEqual({
             type: 'CLEAR_AOI_INFO',
@@ -205,68 +161,6 @@ describe('export actions', () => {
         expect(actions.clearJobInfo()).toEqual({
             type: 'CLEAR_JOB_INFO',
         });
-    });
-
-    it('DrawerTimeout closeDrawer should close drawer', () => {
-        const timeout = new actions.DrawerTimeout();
-        const expectedActions = [
-            { type: actions.types.CLOSING_DRAWER },
-            { type: actions.types.CLOSED_DRAWER },
-        ];
-
-        const store = mockStore({ drawer: 'open' });
-        return store.dispatch(timeout.closeDrawer())
-            .then(() => {
-                expect(store.getActions()).toEqual(expectedActions);
-            });
-    });
-
-    it('DrawerTimeout closeDrawer should close drawer and clear open timeout', () => {
-        const clearStub = sinon.stub(global, 'clearTimeout');
-        const timeout = new actions.DrawerTimeout(null, 'yo');
-        const expectedActions = [
-            { type: actions.types.CLOSING_DRAWER },
-            { type: actions.types.CLOSED_DRAWER },
-        ];
-
-        const store = mockStore({ drawer: 'open' });
-        return store.dispatch(timeout.closeDrawer())
-            .then(() => {
-                expect(store.getActions()).toEqual(expectedActions);
-                expect(clearStub.calledOnce).toBe(true);
-                clearStub.restore();
-            });
-    });
-
-    it('DrawerTimeout openDrawer should open drawer', () => {
-        const timeout = new actions.DrawerTimeout();
-        const expectedActions = [
-            { type: actions.types.OPENING_DRAWER },
-            { type: actions.types.OPENED_DRAWER },
-        ];
-
-        const store = mockStore({ drawer: 'closed' });
-        return store.dispatch(timeout.openDrawer())
-            .then(() => {
-                expect(store.getActions()).toEqual(expectedActions);
-            });
-    });
-
-    it('DrawerTimeout openDrawer should open drawer and clear open timeout', () => {
-        const clearStub = sinon.stub(global, 'clearTimeout');
-        const timeout = new actions.DrawerTimeout('yo', null);
-        const expectedActions = [
-            { type: actions.types.OPENING_DRAWER },
-            { type: actions.types.OPENED_DRAWER },
-        ];
-
-        const store = mockStore({ drawer: 'closed' });
-        return store.dispatch(timeout.openDrawer())
-            .then(() => {
-                expect(store.getActions()).toEqual(expectedActions);
-                expect(clearStub.calledOnce).toBe(true);
-                clearStub.restore();
-            });
     });
 
     it('reRunExport should return a specific run from "api/runs"', () => {
